tests/requirements.txt
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.27
websockets>=12.0
python-dotenv>=1.0
